Initial commitment for the Starling run on m7 for 512k
* This branch targets making Starling runs on ARM CM7 for SPP mode with less than 512KB memory usage

* Updated signal.h to set the MAX_NUM_SAT support on GPS + GAL

* Updated nav meas structure and ephemeris strucutre to use float instead of double to improve memory usage.

+++ include/swiftnav/ephemeris.h
@@ -86,24 +86,24 @@ extern "C" {
 /** Structure containing the GPS ephemeris for one satellite. */
 typedef struct {
   union {
-    double tgd_gps_s;    /**< GPS TGD  */
-    double tgd_qzss_s;   /**< QZSS TGD */
-    double tgd_bds_s[2]; /**< tgd_bds_s[0] = BDS TGD1,
+    float tgd_gps_s;    /**< GPS TGD  */
+    float tgd_qzss_s;   /**< QZSS TGD */
+    float tgd_bds_s[2]; /**< tgd_bds_s[0] = BDS TGD1,
                               tgd_bds_s[1] = BDS TGD2 */
-    double tgd_gal_s[2]; /**< tgd_gal_s[0] = GAL E5a/E1 BGD,
+    float tgd_gal_s[2]; /**< tgd_gal_s[0] = GAL E5a/E1 BGD,
                               tgd_gal_s[1] = GAL E5b/E1 BGD*/
   };
-  double crc;      /**< Amplitude of the cosine harmonic correction term
+  float crc;      /**< Amplitude of the cosine harmonic correction term
                         to the orbit radius [m] */
-  double crs;      /**< Amplitude of the sine harmonic correction term
+  float crs;      /**< Amplitude of the sine harmonic correction term
                         to the orbit radius [m] */
-  double cuc;      /**< Amplitude of the cosine harmonic correction term
+  float cuc;      /**< Amplitude of the cosine harmonic correction term
                         to the argument of latitude [rad] */
-  double cus;      /**< Amplitude of the sine harmonic correction term
+  float cus;      /**< Amplitude of the sine harmonic correction term
                         to the argument of latitude [rad] */
-  double cic;      /**< Amplitude of the cosine harmonic correction term
+  float cic;      /**< Amplitude of the cosine harmonic correction term
                         to the angle of inclination [rad] */
-  double cis;      /**< Amplitude of the sine harmonic correction term
+  float cis;      /**< Amplitude of the sine harmonic correction term
                         to the angle of inclination [rad] */
   double dn;       /**< Mean motion difference from computed value
                         [semi-circles/s] */
@@ -116,9 +116,9 @@ typedef struct {
   double w;        /**< Argument of perigee [semi-circles] */
   double inc;      /**< Inclindation angle at reference time [semi-circles] */
   double inc_dot;  /**< Rate of inclination angle [semi-circles/s] */
-  double af0;      /**< Time offset of the sat clock [s] **/
-  double af1;      /**< Drift of the sat clock [s/s] **/
-  double af2;      /**< Acceleration of the sat clock [s/s^2] **/
+  float af0;      /**< Time offset of the sat clock [s] **/
+  float af1;      /**< Drift of the sat clock [s/s] **/
+  float af2;      /**< Acceleration of the sat clock [s/s^2] **/
   gps_time_t toc;  /**< Reference time of clock. */
   u16 iodc;        /**< Issue of data clock. */
   u16 iode;        /**< Issue of data ephemeris. */
